requirements.txt
opencv-python
deepface
tf-keras
flask
numpy
pillow
twilio
mediapipe
python-dotenv
streamlit
